Complete Phase 4: extraction pipeline (dfl_torch/extract.py, main.py extract_torch)
A clean-room, standalone extraction pipeline -- not a patch to
mainscripts/Extractor.py, consistent with this migration's approach
throughout: parallel implementation, legacy code untouched until
ready to fully retire.

Hit a real architectural blocker: facelib.LandmarksProcessor.
get_transform_mat (DFL's alignment-crop transform) requires dlib's
68-point landmark scheme, but FaceLandmarkDetector returns 478 points
in MediaPipe's own topology -- no shared indexing, and every
downstream consumer (get_transform_mat, get_image_hull_mask,
samplelib.SampleLoader) hard-requires dlib-68. Rather than hand-derive
a correspondence table from memory (wrong indices would silently
misalign every extracted frame, with no real face photo here to catch
it visually), used WebSearch/WebFetch to find and vendor a maintained,
MIT-licensed, purpose-built conversion
(github.com/PeizhiYan/Mediapipe_2_Dlib_Landmarks) as
dfl_torch/alignment.py's convert_mediapipe_landmarks_to_dlib68, with
source/license documented inline.

Found and fixed a second real bug while wiring this up, same class as
the earlier np.int fix: get_transform_mat passes a float64 array to
cv2.getAffineTransform, which this repo's modern OpenCV (5.0.0)
rejects (CV_32F required) -- an older opencv-python accepted the
looser dtype. Fixed with a one-line astype(np.float32) at the call
site, caught by an integration test feeding real converted landmarks
through the actual unstubbed get_transform_mat.

dfl_torch/extract.py: extract_one_image / extract_directory, writing
DFLJPG output with the same fields Extractor.py itself writes, so
output is an ordinary training-ready sample. main.py extract_torch
subcommand wired in (added to the TF-init-skip list alongside
train_torch).

12 new tests, including one that extracted output loads correctly
through the actual samplelib.SampleLoader path
dfl_torch.data.SAEHDFaceDataset uses, not just a structural check.
Verified via a real end-to-end CLI run.

171 tests passing total.

+++ facelib/LandmarksProcessor.py
@@ -346,7 +346,7 @@ def get_transform_mat (image_landmarks, output_size, face_type, scale=1.0):
 
     # calc affine transform from 3 global space points to 3 local space points size of 'output_size'
     pts2 = np.float32(( (0,0),(output_size,0),(output_size,output_size) ))
-    mat = cv2.getAffineTransform(l_t,pts2)
+    mat = cv2.getAffineTransform(l_t.astype(np.float32),pts2)
     return mat
 
 def get_rect_from_landmarks(image_landmarks):
